Adder optimization -- Let verilog compiler synthesize adder.
Instead of gate-level impl of adder, just use add symbol
to implement adder.

It allows verilog compiler synthesize much more faster adder for us
and who knows? if it thinks DSP block is suitable for this kind of
task, it will be able to use it instead of consuming lots of LUTs.